Network V2.0.0
大改！
将原来的网络数据报收发处理从一个类中分离出来，分为两个类：
CommunicateWithClient负责收发数据报
MessageOperate负责处理数据报
